requirements.txt
flask
flask_sqlalchemy
docker
llama-cpp-python
stripe
pillow